Create SetEmptyDates
This is NOT working.
This is running upon a new entry through a Google form. The problem is, that I also populate the associated spreadsheet from a IFTTT recipe. After I insert new entries through IFTTT, the last row is not update for some reason, so when a new entry is inserted form the form, it is not updating the dates for the entry being inserted. 
I am tired enough. 
Next: leave the entry as it is, and run a process every hour that will populate the date and time of all empty cells.